+++ stochastic/__init__.py
@@ -8,15 +8,3 @@ from stochastic._version import __license__
 from stochastic._version import __title__
 from stochastic._version import __url__
 from stochastic._version import __version__
-
-__all__ = (
-    _version,
-    __description__,
-    __url__,
-    __title__,
-    __author__,
-    __author_email__,
-    __license__,
-    __copyright__,
-    __docs_copyright__,
-)