+++ stochastic/processes/diffusion/__init__.py
@@ -1,0 +1,6 @@
+from stochastic.processes.diffusion.constant_elasticity_variance import ConstantElasticityVarianceProcess
+from stochastic.processes.diffusion.cox_ingersoll_ross import CoxIngersollRossProcess
+from stochastic.processes.diffusion.diffusion import DiffusionProcess
+from stochastic.processes.diffusion.extended_vasicek import ExtendedVasicekProcess
+from stochastic.processes.diffusion.ornstein_uhlenbeck import OrnsteinUhlenbeckProcess
+from stochastic.processes.diffusion.vasicek import VasicekProcess