+++ stochastic/processes/discrete/__init__.py
@@ -1,0 +1,5 @@
+from stochastic.processes.discrete.bernoulli import BernoulliProcess
+from stochastic.processes.discrete.chinese_restaurant import ChineseRestaurantProcess
+from stochastic.processes.discrete.markov_chain import MarkovChain
+from stochastic.processes.discrete.moran import MoranProcess
+from stochastic.processes.discrete.random_walk import RandomWalk